babel-rsc: restructure tests
