--- packages/babel-rsc/src/__tests__/test-cases/arrow-1.invalid.jsx
@@ -1,13 +1,0 @@
-export const Test = ({ foo, bar }) => {
-  return (
-    <form
-      action={() => {
-        "use server";
-        console.log("not async", foo, bar);
-      }}
-    >
-      <input name="test" type="text" />
-      <button type="submit">Submit</button>
-    </form>
-  );
-};
