--- packages/babel-rsc/src/__tests__/test-cases/arrow-1.invalid.out.txt
@@ -1,8 +1,0 @@
-arrow-1.invalid.jsx: functions marked with "use server" must be async
-  2 |   return (
-  3 |     <form
-> 4 |       action={() => {
-    |               ^
-  5 |         "use server";
-  6 |         console.log("not async", foo, bar);
-  7 |       }}
