--- packages/babel-rsc/src/__tests__/test-cases/arrow-1.jsx
@@ -1,28 +1,0 @@
-import { doSomethingOnTheServer } from "./server-stuff";
-
-export const Test = ({ foo, bar }) => {
-  return (
-    <form
-      action={async (data) => {
-        "use server";
-        const test = data.get("test");
-        await doSomethingOnTheServer({ test, foo });
-        return { success: true };
-      }}
-    >
-      <input name="test" type="text" />
-      <button type="submit">Submit</button>
-      <button
-        type="button"
-        formAction={async (data) => {
-          "use server";
-          const test = data.get("test");
-          await doSomethingOnTheServer({ test, foo, bar });
-          return { success: true };
-        }}
-      >
-        Submit
-      </button>
-    </form>
-  );
-};
