--- packages/babel-rsc/src/__tests__/test-cases/arrow-2.jsx
@@ -1,16 +1,0 @@
-import { doSomethingOnTheServer } from "./server-stuff";
-
-export const Test = ({ foo }) => {
-  const doStuff = async (data) => {
-    "use server";
-    const test = data.get("test");
-    await doSomethingOnTheServer({ test, foo });
-    return { success: true };
-  };
-  return (
-    <form action={doStuff}>
-      <input name="test" type="text" />
-      <button type="submit">Submit</button>
-    </form>
-  );
-};
