--- packages/babel-rsc/src/__tests__/test-cases/named-2.jsx
@@ -1,17 +1,0 @@
-import { doSomethingOnTheServer } from "./server-stuff";
-
-export const Test = ({ foo }) => {
-  return (
-    <form
-      action={async function doStuff(data) {
-        "use server";
-        const test = data.get("test");
-        await doSomethingOnTheServer({ test, foo });
-        return { success: true };
-      }}
-    >
-      <input name="test" type="text" />
-      <button type="submit">Submit</button>
-    </form>
-  );
-};
