+++ packages/babel-rsc/src/__tests__/test-cases/top-level-directive-default-1/input.jsx
@@ -1,0 +1,8 @@
+"use server";
+import { doSomethingOnTheServer } from "../server-stuff";
+
+const SOME_CONSTANT = "beep";
+
+export default async (formData) => {
+  return doSomethingOnTheServer([formData, SOME_CONSTANT]);
+};
