--- packages/babel-rsc/src/__tests__/test-cases/top-level-directive-default-2.invalid.jsx
@@ -1,8 +1,0 @@
-"use server";
-import { doSomethingOnTheServer } from "./server-stuff";
-
-const SOME_CONSTANT = "beep";
-
-export default async function test(formData) {
-  return doSomethingOnTheServer([formData, SOME_CONSTANT]);
-}
