+++ packages/babel-rsc/src/__tests__/test-cases/top-level-directive-default-2/error.txt
@@ -1,0 +1,8 @@
+input.jsx: Not implemented: 'export default' declarations in "use server" files. Try using 'export { name as default }' instead.
+  4 | const SOME_CONSTANT = "beep";
+  5 |
+> 6 | export default async function test(formData) {
+    | ^
+  7 |   return doSomethingOnTheServer([formData, SOME_CONSTANT]);
+  8 | }
+  9 |
