--- packages/babel-rsc/src/__tests__/test-cases/top-level-directive-default-3.invalid.jsx
@@ -1,10 +1,0 @@
-"use server";
-import { doSomethingOnTheServer } from "./server-stuff";
-
-const SOME_CONSTANT = "beep";
-
-async function test(formData) {
-  return doSomethingOnTheServer([formData, SOME_CONSTANT]);
-}
-
-export default test;
